+++ test/entrypoint/srv_ddl_reload/cartridge_init.lua
@@ -3,6 +3,7 @@
 require('strict').on()
 _G.is_initialized = function() return false end
 
+local fio = require('fio')
 local log = require('log')
 local errors = require('errors')
 local cartridge = require('cartridge')
@@ -14,6 +15,9 @@ if package.setsearchroot ~= nil then
 else
     package.path = package.path .. debug.sourcedir() .. "/?.lua;"
 end
+
+local root = fio.dirname(fio.dirname(fio.dirname(debug.sourcedir())))
+package.path = package.path .. root .. "/?.lua;"
 
 package.preload['customers-storage'] = function()
     local customers_module = {
@@ -34,7 +38,7 @@ package.preload['customers-storage'] = function()
 
     return {
         role_name = 'customers-storage',
-        init = require('storage_init')
+        init = require('storage').init,
     }
 end
