--- test/entrypoint/srv_read_calls_strategies/storage_init.lua
@@ -1,21 +1,0 @@
-return function()
-    if box.info.ro == true then
-        return
-    end
-
-    local engine = os.getenv('ENGINE') or 'memtx'
-    local customers_space = box.schema.space.create('customers', {
-        format = {
-            {name = 'id', type = 'unsigned'},
-            {name = 'bucket_id', type = 'unsigned'},
-            {name = 'name', type = 'string'},
-            {name = 'age', type = 'number'},
-        },
-        if_not_exists = true,
-        engine = engine,
-    })
-    customers_space:create_index('id', {
-        parts = { {field = 'id'} },
-        if_not_exists = true,
-    })
-end
